Move filter bar to folder
