+++ src/vs/workbench/browser/positronDataExplorer/components/dataExplorerPanel/components/filterBar/filterBar.css
